+++ changeo/Receptor.py
@@ -92,6 +92,7 @@ class AIRRSchema:
                                  ('productive', 'functional'),
                                  ('stop_codon', 'stop'),
                                  ('vj_in_frame', 'in_frame'),
+                                 ('locus', 'locus'),
                                  ('v_call', 'v_call'),
                                  ('d_call', 'd_call'),
                                  ('j_call', 'j_call'),
@@ -115,8 +116,7 @@ class AIRRSchema:
     standard_fields = list(_standard_map.keys())
 
     # Custom fields
-    _custom_map = OrderedDict([('locus', 'locus'),
-                               ('c_call', 'c_call'),
+    _custom_map = OrderedDict([('c_call', 'c_call'),
                                ('germline_alignment_d_mask', 'germline_imgt_d_mask'),
                                ('v_score', 'v_score'),
                                ('v_identity', 'v_identity'),
@@ -221,6 +221,7 @@ class ChangeoSchema:
                                  ('STOP', 'stop'),
                                  ('MUTATED_INVARIANT', 'mutated_invariant'),
                                  ('INDELS', 'indels'),
+                                 ('LOCUS', 'locus'),
                                  ('V_CALL', 'v_call'),
                                  ('D_CALL', 'd_call'),
                                  ('J_CALL', 'j_call'),
@@ -279,7 +280,6 @@ class ChangeoSchema:
                                ('P3D_LENGTH', 'p3d_length'),
                                ('P5J_LENGTH', 'p5j_length'),
                                ('D_FRAME', 'd_frame'),
-                               ('LOCUS', 'locus'),
                                ('C_CALL', 'c_call'),
                                ('CDR3_IGBLAST', 'cdr3_igblast'),
                                ('CDR3_IGBLAST_AA', 'cdr3_igblast_aa'),