+++ NEWS.rst
@@ -1,15 +1,21 @@
 Release Notes
 ===============================================================================
 
-Version 0.4.5.999:  April 23, 2019
+Version 0.4.5.999:  May 20, 2019
 -------------------------------------------------------------------------------
 
 DefineClones:
 
-+ Fixed a biopython warning arising when applying `--model aa` to junction
++ Fixed a biopython warning arising when applying ``--model aa`` to junction
   sequences that are not a multiple of three. The junction will now be
   padded with an appropriate number of Ns (usually resulting in a translation
   to X).
+
+MakeDb:
+
++ Combined the extended field arguments of all subcommands (``--scores``,
+  ``--regions``, ``--cdr3``, and ``--junction``) into a single ``--extended``
+  argument.
 
 
 Version 0.4.5:  January 9, 2019